Add DoubleTime, HalfTime and Nightcore support.
