+++ osu.Game/Rulesets/Mods/ModCinema.cs
@@ -1,0 +1,13 @@
+// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using osu.Game.Graphics;
+
+namespace osu.Game.Rulesets.Mods
+{
+    public class ModCinema : ModAutoplay
+    {
+        public override string Name => "Cinema";
+        public override FontAwesome Icon => FontAwesome.fa_osu_mod_cinema;
+    }
+}
