+++ osu.Game/Rulesets/Mods/ModHardRock.cs
@@ -1,0 +1,16 @@
+// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using System;
+using osu.Game.Graphics;
+
+namespace osu.Game.Rulesets.Mods
+{
+    public abstract class ModHardRock : Mod
+    {
+        public override string Name => "Hard Rock";
+        public override FontAwesome Icon => FontAwesome.fa_osu_mod_hardrock;
+        public override string Description => "Everything just got a bit harder...";
+        public override Type[] IncompatibleMods => new[] { typeof(ModEasy) };
+    }
+}
