+++ osu.Game/Rulesets/Mods/ModNightcore.cs
@@ -1,0 +1,25 @@
+// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using osu.Framework.Audio;
+using osu.Framework.Timing;
+using osu.Game.Graphics;
+
+namespace osu.Game.Rulesets.Mods
+{
+    public abstract class ModNightcore : ModDoubleTime
+    {
+        public override string Name => "Nightcore";
+        public override FontAwesome Icon => FontAwesome.fa_osu_mod_nightcore;
+        public override string Description => "uguuuuuuuu";
+
+        public override void ApplyToClock(IAdjustableClock clock)
+        {
+            var pitchAdjust = clock as IHasPitchAdjust;
+            if (pitchAdjust != null)
+                pitchAdjust.PitchAdjust = 1.5;
+            else
+                base.ApplyToClock(clock);
+        }
+    }
+}
