+++ osu.Game/osu.Game.csproj
@@ -109,6 +109,21 @@
     <Compile Include="IO\Serialization\IJsonSerializable.cs" />
     <Compile Include="IPC\ScoreIPCChannel.cs" />
     <Compile Include="Rulesets\BeatmapStatistic.cs" />
+    <Compile Include="Rulesets\Mods\IApplicableToClock.cs" />
+    <Compile Include="Rulesets\Mods\ModAutoplay.cs" />
+    <Compile Include="Rulesets\Mods\ModCinema.cs" />
+    <Compile Include="Rulesets\Mods\ModDoubleTime.cs" />
+    <Compile Include="Rulesets\Mods\ModEasy.cs" />
+    <Compile Include="Rulesets\Mods\ModFlashlight.cs" />
+    <Compile Include="Rulesets\Mods\ModHalfTime.cs" />
+    <Compile Include="Rulesets\Mods\ModHardRock.cs" />
+    <Compile Include="Rulesets\Mods\ModHidden.cs" />
+    <Compile Include="Rulesets\Mods\ModNightcore.cs" />
+    <Compile Include="Rulesets\Mods\ModNoFail.cs" />
+    <Compile Include="Rulesets\Mods\ModPerfect.cs" />
+    <Compile Include="Rulesets\Mods\ModRelax.cs" />
+    <Compile Include="Rulesets\Mods\ModSuddenDeath.cs" />
+    <Compile Include="Rulesets\Mods\MultiMod.cs" />
     <Compile Include="Rulesets\Objects\Legacy\Catch\Hit.cs" />
     <Compile Include="Rulesets\Objects\Legacy\Catch\HitObjectParser.cs" />
     <Compile Include="Rulesets\Objects\Legacy\Catch\Slider.cs" />
